Refactor middleware tests and add separate test files
Consolidated and refactored middleware tests by moving `WithErrorHandling` and `WithThrottler` tests to separate test files (`errors_test.go` and `throttler_test.go`). Simplified test handlers by introducing `HandlerFunc` and updated method calls for consistency. These changes improve test structure and organization, making it more maintainable and readable.
